+++ modules/effects/src/main/java/io/github/palexdev/mfxeffects/animations/ConsumerTransition.java
@@ -18,21 +18,19 @@
 
 package io.github.palexdev.mfxeffects.animations;
 
+import io.github.palexdev.mfxeffects.animations.base.FluentTransition;
 import io.github.palexdev.mfxeffects.enums.Interpolators;
 import javafx.animation.Interpolator;
-import javafx.animation.Transition;
-import javafx.event.ActionEvent;
-import javafx.event.EventHandler;
 import javafx.util.Duration;
 
 import java.util.function.Consumer;
 
 /**
- * A simple implementation of {@link Transition} that allows to specify
+ * A simple implementation of {@link FluentTransition} that allows specifying
  * what to do when the {@link #interpolate(double)} method is called by using
  * a {@link Consumer}.
  */
-public class ConsumerTransition extends Transition {
+public class ConsumerTransition extends FluentTransition {
 	//================================================================================
 	// Properties
 	//================================================================================
@@ -43,51 +41,10 @@ public class ConsumerTransition extends Transition {
 	//================================================================================
 
 	/**
-	 * Sets the transition duration.
-	 */
-	public ConsumerTransition setDuration(Duration duration) {
-		this.setCycleDuration(duration);
-		return this;
-	}
-
-	/**
-	 * Sets the transition duration in milliseconds.
-	 */
-	public ConsumerTransition setDuration(double millis) {
-		this.setCycleDuration(Duration.millis(millis));
-		return this;
-	}
-
-	/**
 	 * Sets the consumer used by the {@link #interpolate(double)} method.
 	 */
 	public ConsumerTransition setInterpolateConsumer(Consumer<Double> interpolateConsumer) {
 		this.interpolateConsumer = interpolateConsumer;
-		return this;
-	}
-
-	/**
-	 * Sets the transition's interpolator.
-	 */
-	public ConsumerTransition setInterpolatorFluent(Interpolator interpolator) {
-		this.setInterpolator(interpolator);
-		return this;
-	}
-
-	public ConsumerTransition setInterpolatorFluent(Interpolators interpolator) {
-		return setInterpolatorFluent(interpolator.toInterpolator());
-	}
-
-	/**
-	 * Sets the transition's delay.
-	 */
-	public ConsumerTransition setDelayFluent(Duration duration) {
-		this.setDelay(duration);
-		return this;
-	}
-
-	public ConsumerTransition setOnFinishedFluent(EventHandler<ActionEvent> handler) {
-		setOnFinished(handler);
 		return this;
 	}
 
@@ -127,42 +84,42 @@ public class ConsumerTransition extends Transition {
 	/**
 	 * Creates a new {@code ConsumerTransition} with the given consumer and duration.
 	 */
-	public static ConsumerTransition of(Consumer<Double> interpolateConsumer, Duration duration) {
+	public static FluentTransition of(Consumer<Double> interpolateConsumer, Duration duration) {
 		return (new ConsumerTransition()).setInterpolateConsumer(interpolateConsumer).setDuration(duration);
 	}
 
 	/**
 	 * Creates a new {@code ConsumerTransition} with the given consumer and duration in milliseconds.
 	 */
-	public static ConsumerTransition of(Consumer<Double> interpolateConsumer, double duration) {
+	public static FluentTransition of(Consumer<Double> interpolateConsumer, double duration) {
 		return (new ConsumerTransition()).setInterpolateConsumer(interpolateConsumer).setDuration(duration);
 	}
 
 	/**
 	 * Creates a new {@code ConsumerTransition} with the given consumer, duration and interpolator.
 	 */
-	public static ConsumerTransition of(Consumer<Double> interpolateConsumer, Duration duration, Interpolator interpolator) {
+	public static FluentTransition of(Consumer<Double> interpolateConsumer, Duration duration, Interpolator interpolator) {
 		return (new ConsumerTransition()).setInterpolateConsumer(interpolateConsumer).setDuration(duration).setInterpolatorFluent(interpolator);
 	}
 
 	/**
 	 * Creates a new {@code ConsumerTransition} with the given consumer, duration in milliseconds and interpolator.
 	 */
-	public static ConsumerTransition of(Consumer<Double> interpolateConsumer, double duration, Interpolator interpolator) {
+	public static FluentTransition of(Consumer<Double> interpolateConsumer, double duration, Interpolator interpolator) {
 		return (new ConsumerTransition()).setInterpolateConsumer(interpolateConsumer).setDuration(duration).setInterpolatorFluent(interpolator);
 	}
 
 	/**
 	 * Creates a new {@code ConsumerTransition} with the given consumer, duration and interpolator.
 	 */
-	public static ConsumerTransition of(Consumer<Double> interpolateConsumer, Duration duration, Interpolators interpolator) {
+	public static FluentTransition of(Consumer<Double> interpolateConsumer, Duration duration, Interpolators interpolator) {
 		return (new ConsumerTransition()).setInterpolateConsumer(interpolateConsumer).setDuration(duration).setInterpolatorFluent(interpolator.toInterpolator());
 	}
 
 	/**
 	 * Creates a new {@code ConsumerTransition} with the given consumer, duration in milliseconds and interpolator.
 	 */
-	public static ConsumerTransition of(Consumer<Double> interpolateConsumer, double duration, Interpolators interpolator) {
+	public static FluentTransition of(Consumer<Double> interpolateConsumer, double duration, Interpolators interpolator) {
 		return (new ConsumerTransition()).setInterpolateConsumer(interpolateConsumer).setDuration(duration).setInterpolatorFluent(interpolator.toInterpolator());
 	}
 }