+++ modules/release/CHANGELOG.md
@@ -16,6 +16,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 [//]: ##[Unreleased]
 
+## [11.24.1] - 19-10-2023
+
+- Bump components and release modules to version 11.24.1
+- Bump effects module to version 11.3.1
+
 ## [11.24.0] - 04-10-2023
 
 - Bump components and release modules to version 11.24.0